net: Phase 3 - Eliminate SmoltcpDevice, integrate poll into LinkLayer
Remove the SmoltcpDevice bridge enum by adding create_smoltcp_iface()
and poll_smoltcp() directly to the LinkLayer trait. Each concrete
link type (LoopbackLink, VirtioLink) now handles its own smoltcp
poll cycle using its private Device impl.

NetIface no longer holds a SmoltcpDevice; poll dispatch goes through
link.write().poll_smoltcp(timestamp, iface, sockets).
